feat: Consume `SchoolStatus` when rendering School spending/comparison views
