+++ web/src/Web.App/Domain/Establishment/SchoolStatus.cs
@@ -1,0 +1,9 @@
+namespace Web.App.Domain;
+
+public class SchoolStatus
+{
+    public string? URN { get; set; }
+    public string? SchoolName { get; set; }
+    public bool IsPartOfTrust { get; set; }
+    public bool IsMat { get; set; }
+}
