+++ web/src/Web.App/Views/SchoolSpending/CustomData.cshtml
@@ -33,6 +33,7 @@
         urn = Model.Urn,
         isCustomData = true,
         isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat,
         resources = Model.Resources
     })
 
@@ -43,6 +44,7 @@
         urn = Model.Urn,
         isCustomData = true,
         isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat,
         resources = Model.Resources
     })
 }
@@ -62,7 +64,8 @@
         id = "low-priority",
         urn = Model.Urn,
         isCustomData = true,
-        isPartOfTrust = Model.IsPartOfTrust
+        isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat
     })
 }
