+++ web/src/Web.App/Views/SchoolSpending/CustomDataSsr.cshtml
@@ -34,6 +34,7 @@
         urn = Model.Urn,
         isCustomData = true,
         isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat,
         resources = Model.Resources
     })
 
@@ -44,6 +45,7 @@
         urn = Model.Urn,
         isCustomData = true,
         isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat,
         resources = Model.Resources
     })
 }
@@ -64,7 +66,7 @@
         urn = Model.Urn,
         isCustomData = true,
         isPartOfTrust = Model.IsPartOfTrust,
-        resources = Model.Resources
+        isMat = Model.IsMat
     })
 }
