+++ web/src/Web.App/Views/SchoolSpending/IndexSsr.cshtml
@@ -58,6 +58,7 @@
         id = "high-priority",
         urn = Model.Urn,
         isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat,
         resources = Model.Resources
     })
 
@@ -67,6 +68,7 @@
         id = "medium-priority",
         urn = Model.Urn,
         isPartOfTrust = Model.IsPartOfTrust,
+        isMat = Model.IsMat,
         resources = Model.Resources
     })
 }
@@ -86,7 +88,7 @@
         id = "low-priority",
         urn = Model.Urn,
         isPartOfTrust = Model.IsPartOfTrust,
-        resources = Model.Resources
+        isMat = Model.IsMat
     })
 }
